requirements.txt
colorama
fastapi
gradio>=3.50.2
keybert
matplotlib
nltk
numpy
openpyxl
pandas
pgvector
pillow
psutil
psycopg2-binary
requests
scikit-learn
scipy
seaborn
sentence-transformers>=2.2.2
spacy
sqlalchemy
tabulate
tensorflow
tf-keras
tqdm
transformers
uvicorn
pillow
tf-keras
psycopg2-binary
requests
tabulate
colorama
pgvector
psutil
pydantic-settings